Start page, game on page, result page
I have worked on the start page and next step is to make the "start game" a button - so you can click on that to come to the "game on page". Right now you can click everywhere on the screen in order to come to the game on page - which isn't the purpose.